# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List all products
      description: Returns all products in the catalog
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"

    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product
      description: Add a new product to the catalog (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "400":
          description: Invalid input
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Product found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Product not found

    put:
      operationId: updateProduct
      tags: [products]
      summary: Update a product
      description: Update an existing product (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Product updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "400":
          description: Invalid input
        "404":
          description: Product not found

    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product
      description: Remove a product from the catalog (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Product deleted
        "404":
          description: Product not found
  /products/{id}/sold:
    post:
      operationId: recordProductSale
      tags: [products]
      summary: Adjust sold count for a product (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Sold count updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Product not found
...
  /admin/products/{id}/published:
    patch:
      operationId: setProductPublished
      tags: [products]
      summary: Toggle product published state (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetProductPublished"
      responses:
        "200":
          description: Product updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Product not found
...
components:
  schemas:
...
    PricingOption:
      type: object
      required: [duration, price]
      properties:
        duration:
          type: string
        price:
          type: number
          minimum: 0
        salePrice:
          type: ["number", "null"]
        priceUsd:
          type: ["number", "null"]
          minimum: 0
        salePriceUsd:
          type: ["number", "null"]
          minimum: 0
    Product:
      type: object
      required:
        - id
        - name
        - price
        - duration
        - createdAt
      properties:
        id:
          type: number
        name:
          type: string
        category:
          type: ["string", "null"]
        brand:
          type: ["string", "null"]
        coverImageUrl:
          type: ["string", "null"]
        price:
          type: number
        salePrice:
          type: ["number", "null"]
        priceUsd:
          type: ["number", "null"]
        salePriceUsd:
          type: ["number", "null"]
        pricingOptions:
          type: ["array", "null"]
          items:
            $ref: '#/components/schemas/PricingOption'
        duration:
          type: string
        deliveryTime:
          type: ["string", "null"]
        activationType:
          type: ["string", "null"]
        onCustomerAccount:
          type: ["boolean", "null"]
        invitationLink:
          type: ["string", "null"]
        licenseKey:
          type: ["string", "null"]
        sharedAccount:
          type: ["boolean", "null"]
        description:
          type: ["string", "null"]
        features:
          type: ["array", "null"]
          items:
            type: string
        warrantyDuration:
          type: ["string", "null"]
        customerInfoRequired:
          type: ["array", "null"]
          items:
            type: string
        afterPurchaseInstructions:
          type: ["string", "null"]
        soldCount:
          type: number
        published:
          type: boolean
        createdAt:
          type: string
    ProductInput:
      type: object
      required:
        - name
        - price
        - duration
      properties:
        name:
          type: string
          minLength: 1
        category:
          type: string
        brand:
          type: string
        coverImageUrl:
          type: string
        price:
          type: number
          minimum: 0
        salePrice:
          type: number
          minimum: 0
        priceUsd:
          type: number
          minimum: 0
        salePriceUsd:
          type: number
          minimum: 0
        pricingOptions:
          type: array
          items:
            $ref: '#/components/schemas/PricingOption'
        duration:
          type: string
          minLength: 1
        deliveryTime:
          type: string
        activationType:
          type: string
        onCustomerAccount:
          type: boolean
        invitationLink:
          type: string
        licenseKey:
          type: string
        sharedAccount:
          type: boolean
        description:
          type: string
        features:
          type: array
          items:
            type: string
        warrantyDuration:
          type: string
        customerInfoRequired:
          type: array
          items:
            type: string
        afterPurchaseInstructions:
          type: string
...
    ProductUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        category:
          type: string
        brand:
          type: string
        coverImageUrl:
          type: string
        price:
          type: number
          minimum: 0
        salePrice:
          type: number
          minimum: 0
        priceUsd:
          type: number
          minimum: 0
        salePriceUsd:
          type: number
          minimum: 0
        pricingOptions:
          type: array
          items:
            $ref: '#/components/schemas/PricingOption'
        duration:
          type: string
          minLength: 1
        deliveryTime:
          type: string
        activationType:
          type: string
        onCustomerAccount:
          type: boolean
        invitationLink:
          type: string
        licenseKey:
          type: string
        sharedAccount:
          type: boolean
        description:
          type: string
        features:
          type: array
          items:
            type: string
        warrantyDuration:
          type: string
        customerInfoRequired:
          type: array
          items:
            type: string
        afterPurchaseInstructions:
          type: string
        published:
          type: boolean
    SetProductPublished:
      type: object
      required: [published]
      properties:
        published:
          type: boolean
...