# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PricingOption } from './pricingOption';

export interface Product {
  id: number;
  name: string;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  brand?: string | null;
  /** @nullable */
  coverImageUrl?: string | null;
  price: number;
  /** @nullable */
  salePrice?: number | null;
  /** @nullable */
  priceUsd?: number | null;
  /** @nullable */
  salePriceUsd?: number | null;
  /** @nullable */
  pricingOptions?: PricingOption[] | null;
  duration: string;
  /** @nullable */
  deliveryTime?: string | null;
  /** @nullable */
  activationType?: string | null;
  /** @nullable */
  onCustomerAccount?: boolean | null;
  /** @nullable */
  invitationLink?: string | null;
  /** @nullable */
  licenseKey?: string | null;
  /** @nullable */
  sharedAccount?: boolean | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  features?: string[] | null;
  /** @nullable */
  warrantyDuration?: string | null;
  /** @nullable */
  customerInfoRequired?: string[] | null;
  /** @nullable */
  afterPurchaseInstructions?: string | null;
  soldCount?: number;
  published?: boolean;
  createdAt: string;
}