# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PricingOption } from './pricingOption';

export interface ProductInput {
  /** @minLength 1 */
  name: string;
  category?: string;
  brand?: string;
  coverImageUrl?: string;
  /** @minimum 0 */
  price: number;
  /** @minimum 0 */
  salePrice?: number;
  /** @minimum 0 */
  priceUsd?: number;
  /** @minimum 0 */
  salePriceUsd?: number;
  pricingOptions?: PricingOption[];
  /** @minLength 1 */
  duration: string;
  deliveryTime?: string;
  activationType?: string;
  onCustomerAccount?: boolean;
  invitationLink?: string;
  licenseKey?: string;
  sharedAccount?: boolean;
  description?: string;
  features?: string[];
  warrantyDuration?: string;
  customerInfoRequired?: string[];
  afterPurchaseInstructions?: string;
}